+++ OpenGL Smiley Editor/OpenGLApplication/BrandonUtils.h
@@ -20,11 +20,11 @@ namespace BrandonUtils
     float map(float current, float oldMin, float oldMax, float newMin, float newMax);
 
     /// <summary>
-    /// Checks if both floats are "close enough" to be equal. The value of "close enough" depends on the value of the floats.
+    /// Checks if both floating point values are "close enough" to be equal. The value of "close enough" depends on the value of the floats.
     /// If the floats are huge, "close enough" will be big. And if the floats are small, "close enough" will be very small.
     /// This is Donald Knuth's algorithm on relative epsilons.
     /// </summary>
-    bool isApproximatelyEqual(float first, float second);
+    bool isApproximatelyEqual(double first, double second);
 
     /// <summary>
     /// Converts the given vector from window coordinates to OpenGL coordinates, with a specified window size.
